Add support EMW3080B

+++ include/mico_platform.h
@@ -52,16 +52,16 @@ typedef platform_spi_slave_command_t            mico_spi_slave_command_t;
 
 typedef platform_spi_slave_data_buffer_t        mico_spi_slave_data_buffer_t;
 
-typedef int32_t mico_gpio_t;  /**< MiCO GPIO peripheral handle, MICO_GPIO_XX define by board/<board_name>/board.h. */
-typedef int32_t mico_spi_t;   /**< MiCO SPI peripheral handle, MICO_SPI_XX define by board/<board_name>/board.h. */
-typedef int32_t mico_iis_t;   /**< MiCO IIS peripheral handle, MICO_IIS_XX define by board/<board_name>/board.h. */
-typedef int32_t mico_adc_t;   /**< MiCO ADC peripheral handle, MICO_ADC_XX define by board/<board_name>/board.h. */
-typedef int32_t mico_i2c_t;   /**< MiCO I2C peripheral handle, MICO_I2C_XX define by board/<board_name>/board.h. */
-typedef int32_t mico_pwm_t;   /**< MiCO PWM peripheral handle, MICO_PWM_XX define by board/<board_name>/board.h. */
-typedef int32_t mico_uart_t;  /**< MiCO UART peripheral handle, MICO_UART_XX define by board/<board_name>/board.h. */
+typedef int8_t mico_gpio_t;  /**< MiCO GPIO peripheral handle, MICO_GPIO_XX define by board/<board_name>/board.h. */
+typedef int8_t mico_spi_t;   /**< MiCO SPI peripheral handle, MICO_SPI_XX define by board/<board_name>/board.h. */
+typedef int8_t mico_iis_t;   /**< MiCO IIS peripheral handle, MICO_IIS_XX define by board/<board_name>/board.h. */
+typedef int8_t mico_adc_t;   /**< MiCO ADC peripheral handle, MICO_ADC_XX define by board/<board_name>/board.h. */
+typedef int8_t mico_i2c_t;   /**< MiCO I2C peripheral handle, MICO_I2C_XX define by board/<board_name>/board.h. */
+typedef int8_t mico_pwm_t;   /**< MiCO PWM peripheral handle, MICO_PWM_XX define by board/<board_name>/board.h. */
+typedef int8_t mico_uart_t;  /**< MiCO UART peripheral handle, MICO_UART_XX define by board/<board_name>/board.h. */
 
-typedef int32_t mico_flash_t;
-typedef int32_t mico_partition_t;
+typedef int8_t mico_flash_t;
+typedef int8_t mico_partition_t;
 
 
 #include "MiCODrivers/mico_gpio.h"
